requirements.txt
# Editable install with no version control (api==0.0.0)
-e /home/recruit/dev/api_project
certifi==2019.11.28
chardet==3.0.4
Deprecated==1.2.7
idna==2.9
PyGithub==1.47
PyJWT==1.7.1
requests==2.23.0
urllib3==1.25.8
wrapt==1.12.1
